+++ CHANGELOG.md
@@ -12,9 +12,11 @@ The format of the release notes follows [Keep a Changelog](https://keepachangelo
   download sessions ([#16])
 - Method `clean` to clean the product folder recursively from any files not belonging to the inventory
   ([#16])
-- Methods `ignore!` and `unignore!` to exclude database files in the inventory and include
+- Functions `ignore!` and `unignore!` to exclude database files in the inventory and include
   them again ([#17])
-- Methods `attach!` and `detach!` to mark extra data exempt from `clean` procedures ([#17])
+- Functions `attach!` and `detach!` to mark extra data exempt from `clean` procedures ([#17])
+- Function `list_inventory` to show a simplified folder and file tree and statistics about
+  the inventory and the downloaded portion ([#17])
 
 ### Changed